+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Change Log
 - [ADDED] multiple costs check in create functions and runopp
 - [ADDED] correct_dtypes() function for fileIO convert
 - [FIXED] revise to_ppc() and to_mpc() init behaviour
+- [CHANGED] import requirements / dependencies
 
 [2.7.1]- 2021-07-22
 ----------------------